Move PathResolver data structures to a separate module
