+++ src/Juvix/Compiler/Concrete/Translation/FromParsed/Analysis/PathResolver.hs
@@ -1,22 +1,13 @@
 module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver
   ( module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Paths,
+    module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Base,
     module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Error,
+    module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Data,
     module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.PackageInfo,
-    module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.DependenciesConfig,
-    PathResolver (..),
-    registerDependencies,
-    withPath,
-    withPathFile,
-    expectedModulePath,
     runPathResolver,
     runPathResolverPipe,
     runPathResolverPipe',
     evalPathResolverPipe,
-    resolverCache,
-    resolverCacheItemPackage,
-    ResolverState,
-    resolverFiles,
-    iniResolverState,
   )
 where
 
@@ -24,7 +15,8 @@ import Data.HashMap.Strict qualified as HashMap
 import Data.HashSet qualified as HashSet
 import Data.Text qualified as T
 import Juvix.Compiler.Concrete.Data.Name
-import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.DependenciesConfig
+import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Base
+import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Data
 import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Error
 import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.PackageInfo
 import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Paths
@@ -35,67 +27,6 @@ import Juvix.Data.Effect.Git
 import Juvix.Extra.Paths
 import Juvix.Extra.Stdlib (ensureStdlib)
 import Juvix.Prelude
-
-data PathResolver m a where
-  RegisterDependencies :: DependenciesConfig -> PathResolver m ()
-  ExpectedModulePath :: Path Abs File -> TopModulePath -> PathResolver m (Maybe (Path Abs File))
-  WithPath ::
-    TopModulePath ->
-    (Either PathResolverError (Path Abs Dir, Path Rel File) -> m x) ->
-    PathResolver m x
-
-makeSem ''PathResolver
-
-data ResolverEnv = ResolverEnv
-  { _envRoot :: Path Abs Dir,
-    -- | The path of the input file *if* we are using the global project
-    _envSingleFile :: Maybe (Path Abs File),
-    _envLockfileInfo :: Maybe LockfileInfo
-  }
-
-data ResolverCacheItem = ResolverCacheItem
-  { _resolverCacheItemPackage :: PackageInfo,
-    _resolverCacheItemDependency :: LockfileDependency
-  }
-  deriving stock (Show)
-
-data ResolverState = ResolverState
-  { -- | juvix files indexed by relative path
-    _resolverFiles :: HashMap (Path Rel File) (NonEmpty PackageInfo),
-    -- | PackageInfos indexed by root
-    _resolverCache :: HashMap (Path Abs Dir) ResolverCacheItem,
-    _resolverShouldWriteLockfile :: Bool
-  }
-  deriving stock (Show)
-
-data ResolvedDependency = ResolvedDependency
-  { _resolvedDependencyPath :: Path Abs Dir,
-    _resolvedDependencyDependency :: Dependency
-  }
-
-makeLenses ''ResolverState
-makeLenses ''ResolverEnv
-makeLenses ''ResolvedDependency
-makeLenses ''ResolverCacheItem
-
-iniResolverState :: ResolverState
-iniResolverState =
-  ResolverState
-    { _resolverCache = mempty,
-      _resolverFiles = mempty,
-      _resolverShouldWriteLockfile = False
-    }
-
-checkShouldWriteLockfile :: (Member (State ResolverState) r) => ResolvedDependency -> Sem r ()
-checkShouldWriteLockfile d = case d ^. resolvedDependencyDependency of
-  DependencyGit {} -> modify' (set resolverShouldWriteLockfile True)
-  DependencyPath {} -> return ()
-
-withEnvRoot :: (Members '[Reader ResolverEnv] r) => Path Abs Dir -> Sem r a -> Sem r a
-withEnvRoot root' = local (set envRoot root')
-
-withLockfile :: (Members '[Reader ResolverEnv] r) => LockfileInfo -> Sem r a -> Sem r a
-withLockfile f = local (set envLockfileInfo (Just f))
 
 mkPackage ::
   forall r.
@@ -187,9 +118,6 @@ mkPackageInfo mpackageEntry _packageRoot pkg = do
 
 lookupCachedDependency :: (Members '[State ResolverState, Reader ResolverEnv, Files, GitClone] r) => Path Abs Dir -> Sem r (Maybe LockfileDependency)
 lookupCachedDependency p = fmap (^. resolverCacheItemDependency) . HashMap.lookup p <$> gets (^. resolverCache)
-
-withPathFile :: (Members '[PathResolver] r) => TopModulePath -> (Either PathResolverError (Path Abs File) -> Sem r a) -> Sem r a
-withPathFile m f = withPath m (f . mapRight (uncurry (<//>)))
 
 resolveDependency :: forall r. (Members '[Reader ResolverEnv, Files, Error DependencyError, GitClone] r) => PackageDependencyInfo -> Sem r ResolvedDependency
 resolveDependency i = case i ^. packageDepdendencyInfoDependency of
