+++ src/Juvix/Compiler/Concrete/Translation/FromParsed/Analysis/PathResolver/Base.hs
@@ -1,0 +1,23 @@
+module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Base
+  ( module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Base,
+    module Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.DependenciesConfig,
+  )
+where
+
+import Juvix.Compiler.Concrete.Data.Name
+import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.DependenciesConfig
+import Juvix.Compiler.Concrete.Translation.FromParsed.Analysis.PathResolver.Error
+import Juvix.Prelude
+
+data PathResolver m a where
+  RegisterDependencies :: DependenciesConfig -> PathResolver m ()
+  ExpectedModulePath :: Path Abs File -> TopModulePath -> PathResolver m (Maybe (Path Abs File))
+  WithPath ::
+    TopModulePath ->
+    (Either PathResolverError (Path Abs Dir, Path Rel File) -> m x) ->
+    PathResolver m x
+
+makeSem ''PathResolver
+
+withPathFile :: (Members '[PathResolver] r) => TopModulePath -> (Either PathResolverError (Path Abs File) -> Sem r a) -> Sem r a
+withPathFile m f = withPath m (f . mapRight (uncurry (<//>)))
